Adopt Scala3 syntax

+++ modules/server/src/main/scala/scaladex/server/service/DependencyUpdater.scala
@@ -4,43 +4,42 @@ import scala.concurrent.ExecutionContext
 import scala.concurrent.Future
 import scala.util.control.NonFatal
 
-import com.typesafe.scalalogging.LazyLogging
 import scaladex.core.model.Project
 import scaladex.core.service.ProjectService
 import scaladex.core.service.SchedulerDatabase
-import scaladex.core.util.ScalaExtensions._
+import scaladex.core.util.ScalaExtensions.*
 
-class DependencyUpdater(database: SchedulerDatabase, projectService: ProjectService)(implicit ec: ExecutionContext)
-    extends LazyLogging {
+import com.typesafe.scalalogging.LazyLogging
+
+class DependencyUpdater(database: SchedulerDatabase, projectService: ProjectService)(using ExecutionContext)
+    extends LazyLogging:
 
   def updateAll(): Future[String] =
-    for {
-      status <- updateProjectDependencyTable()
-    } yield status
+    for status <- updateProjectDependencyTable()
+    yield status
 
   def updateProjectDependencyTable(): Future[String] =
-    for {
+    for
       allProjects <- database.getAllProjects()
       _ = logger.info(s"Updating dependencies of ${allProjects.size} projects")
       _ <- allProjects.mapSync(updateDependencies)
-    } yield s"Updated dependencies of ${allProjects.size} projects"
+    yield s"Updated dependencies of ${allProjects.size} projects"
 
-  def updateDependencies(project: Project): Future[Unit] = {
+  def updateDependencies(project: Project): Future[Unit] =
     val future =
-      if (project.githubStatus.isMoved)
-        database.deleteProjectDependencies(project.reference).map(_ => ())
+      if project.githubStatus.isMoved then database.deleteProjectDependencies(project.reference).map(_ => ())
       else
-        for {
+        for
           header <- projectService.getHeader(project)
           dependencies <- header
             .map(h => database.computeProjectDependencies(project.reference, h.latestVersion))
             .getOrElse(Future.successful(Seq.empty))
           _ <- database.deleteProjectDependencies(project.reference)
           _ <- database.insertProjectDependencies(dependencies)
-        } yield ()
+        yield ()
     future.recover {
       case NonFatal(cause) =>
         logger.error(s"Failed to update dependencies of ${project.reference} of status ${project.githubStatus}", cause)
     }
-  }
-}
+  end updateDependencies
+end DependencyUpdater
